refactor: extract the settings tab, settings types, and modals
Second extraction in the main.ts split, and the biggest. main.ts goes from
5,684 lines to 3,151.

Three new files:

- settings-types.ts (230): PlaudImporterSettings, DEFAULT_SETTINGS,
  CURRENT_SETTINGS_VERSION, the ribbon-icon list and resolveRibbonIconId. Its
  own file because three places need it and none should have to import the
  plugin to get it: the plugin, the settings tab, and the capture store's
  generic parameter.
- settings-tab.ts (2,167): the tab plus the 36 copy constants that are
  tab-only, behind a SettingsTabHost interface naming the 19 plugin members
  the tab actually uses. The interface extends Plugin because
  PluginSettingTab's constructor takes a real one, not a structural stand-in.
- modals.ts (220): all three modals. Pulled forward from its own planned PR
  because the tab uses BrowserSignInModal, so leaving them in main.ts would
  have needed a circular import.

ReauthOutcome moves to capture-store.ts, next to CaptureStoreResult: the two
are the same kind of type and its doc comment already explains itself in
terms of that one.

Both moved bodies are BYTE-IDENTICAL to their originals, verified by diff:
the tab after renaming PlaudImporterPlugin to SettingsTabHost, the modals
apart from the export keyword.

Two things this turned up:

scripts/check-icons.mjs read main.ts and nothing else, so moving
RIBBON_ICON_CHOICES took the ribbon icons out of its view entirely. It caught
itself, which is that guard's "could not locate the array" branch earning its
keep. It now scans every top-level .ts, so the next extraction cannot
silently narrow it. Re-verified that it still fails on a bad icon id.

The previous commit deleted the comment above SIGN_IN_NOTE, which was fused
to the deep-link comment block that moved to capture-store.ts. Restored.

No behavior change. 1,211 tests pass, unchanged.
